+++ AoC2024/Day22/Day22.h
@@ -21,6 +21,6 @@ namespace AoC2024 {
         }
 
     private:
-        std::vector<std::uint64_t> input = {};
+        std::vector<std::uint32_t> input = {};
     };
 }
